+++ apps/ngx-material-intl-tel-input/src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, signal } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { NgxMaterialIntlTelInputComponent } from 'ngx-material-intl-tel-input';
 
@@ -8,7 +8,13 @@ import { NgxMaterialIntlTelInputComponent } from 'ngx-material-intl-tel-input';
   selector: 'ngx-material-intl-tel-input-root',
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AppComponent {
   title = 'ngx-material-intl-tel-input';
+  currentPhoneValue = signal<string>('');
+
+  getValue(value: string): void {
+    this.currentPhoneValue.set(value);
+  }
 }