+++ apps/ngx-material-intl-tel-input/src/app/app.config.ts
@@ -1,8 +1,20 @@
-import { ApplicationConfig } from '@angular/core';
+import {
+  ApplicationConfig,
+  provideExperimentalZonelessChangeDetection
+} from '@angular/core';
 import { provideRouter } from '@angular/router';
 import { appRoutes } from './app.routes';
 import { provideAnimations } from '@angular/platform-browser/animations';
+import {
+  provideHttpClient,
+  withInterceptorsFromDi
+} from '@angular/common/http';
 
 export const appConfig: ApplicationConfig = {
-  providers: [provideRouter(appRoutes), provideAnimations()],
+  providers: [
+    provideRouter(appRoutes),
+    provideAnimations(),
+    provideHttpClient(withInterceptorsFromDi()),
+    provideExperimentalZonelessChangeDetection()
+  ]
 };